+++ src/lc-lib/admin/responses.go
@@ -55,5 +55,8 @@ func init() {
 
   // SNAP
   gob.Register([]*core.Snapshot{&core.Snapshot{}})
+  // SNAP - time.Time
   gob.Register(time.Now())
+  // SNAP - time.Duration
+  gob.Register(time.Since(time.Now()))
 }